# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "urtorrentd",
    "description": "HTTP API of urtorrentd, a BitTorrent daemon on the urtorrent library. Units: bytes, bytes per second, seconds, unix seconds; `null` means unknown or unlimited. Errors always have the `ErrorBody` shape with a stable `code`.",
    "contact": {
      "name": "urtorrentd contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/torrents/{hash}": {
      "get": {
        "tags": [
          "torrent"
        ],
        "summary": "Everything about one torrent.",
        "operationId": "get_torrent",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentDetail"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "torrent"
        ],
        "summary": "Change the display name or comment.",
        "operationId": "patch_torrent",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TorrentPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Changed."
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/{hash}/files": {
      "get": {
        "tags": [
          "torrent"
        ],
        "summary": "The content files.",
        "operationId": "list_files",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileInfo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/{hash}/files/priority": {
      "post": {
        "tags": [
          "torrent"
        ],
        "summary": "Set the priority of files (0 skips them).",
        "operationId": "set_file_priority",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilePriorityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Changed."
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/{hash}/files/rename": {
      "post": {
        "tags": [
          "torrent"
        ],
        "summary": "Rename (move) a file within the save path.",
        "operationId": "rename_file",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Renamed."
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/{hash}/folders/rename": {
      "post": {
        "tags": [
          "torrent"
        ],
        "summary": "Rename a folder (every file under it moves).",
        "operationId": "rename_folder",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Renamed."
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/{hash}/peers": {
      "get": {
        "tags": [
          "torrent"
        ],
        "summary": "Connected peers.",
        "operationId": "list_peers",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeerInfo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/torrents/{hash}/pieces": {
      "get": {
        "tags": [
          "torrent"
        ],
        "summary": "Every piece's state and availability.",
        "operationId": "get_pieces",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Info-hash, 40 hex characters.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PiecesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request or invalid value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Refused: banned address, foreign origin or host not allowed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such torrent or resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Disk or internal failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Shutting down.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "Every error response has this body.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error."
          }
        }
      },
...
      "FilePriorityRequest": {
        "type": "object",
        "description": "Set file priorities.",
        "required": [
          "indexes",
          "priority"
        ],
        "properties": {
          "indexes": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 0
            },
            "description": "File indexes."
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "description": "0 = skip, 1..=7, higher first; 4 is normal.",
            "maximum": 7,
            "minimum": 0
          }
        }
      },
...
      "RenameRequest": {
        "type": "object",
        "description": "Rename a file or folder inside a torrent.",
        "required": [
          "old_path",
          "new_path"
        ],
        "properties": {
          "old_path": {
            "type": "string",
            "description": "Current path relative to the save path."
          },
          "new_path": {
            "type": "string",
            "description": "New path relative to the save path."
          }
        }
      },
...
      "TorrentDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TorrentSummary",
            "description": "The list row."
          },
          {
            "type": "object",
            "required": [
              "web_seeds",
              "known_peers",
              "source_url"
            ],
            "properties": {
              "web_seeds": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Web seed URLs."
              },
              "known_peers": {
                "type": "integer",
                "description": "Peer addresses known (connected or waiting to be dialled).",
                "minimum": 0
              },
              "source_url": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "The URL the torrent was added from."
              }
            }
          }
        ],
        "description": "Everything about one torrent."
      },
...
      "TorrentPatch": {
        "type": "object",
        "description": "Change a torrent's display name or comment. Absent fields stay; `null`\nrestores the torrent's own value.",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name."
          },
          "comment": {
            "type": [
              "string",
              "null"
            ],
            "description": "Comment."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key from POST /api/v1/auth/api-key."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "urtorrentd_sid",
        "description": "Set by POST /api/v1/auth/login."
      }
    }
  },
  "security": [
    {
      "session": []
    },
    {
      "api_key": []
    }
  ],
...
}